refactor(command): split internal command components
